Setting up the server side calls for todo and fixing a filter bug
I set up the server with all the appropriate calls and types for todo's so it is displayed on the page. I fixed a problem in my filter by owner method, I was using a double equals rather than .equals which was causing me to return zero todo's and therefore there was nothing to filter.